Refactor combat math into helper
