+++ combat/engine/__init__.py
@@ -4,6 +4,7 @@ from .common import _current_hp, CombatParticipant
 from .turn_manager import TurnManager
 from .aggro_tracker import AggroTracker
 from .damage_processor import DamageProcessor
+from .combat_math import CombatMath
 from .combat_engine import CombatEngine
 
 __all__ = [
@@ -11,6 +12,7 @@ __all__ = [
     "TurnManager",
     "AggroTracker",
     "DamageProcessor",
+    "CombatMath",
     "CombatParticipant",
     "_current_hp",
 ]
