+++ typeclasses/tests/test_unarmed_damage_bonus.py
@@ -57,9 +57,9 @@ class TestUnarmedAutoAttack(unittest.TestCase):
              patch("world.system.state_manager.apply_regen"), \
              patch("world.system.state_manager.get_effective_stat", return_value=0), \
              patch("world.system.stat_manager.check_hit", return_value=True), \
-             patch("combat.engine.utils.roll_evade", return_value=False), \
-             patch("combat.engine.utils.roll_parry", return_value=False), \
-             patch("combat.engine.utils.roll_block", return_value=False), \
+             patch("combat.engine.combat_math.roll_evade", return_value=False), \
+             patch("combat.engine.combat_math.roll_parry", return_value=False), \
+             patch("combat.engine.combat_math.roll_block", return_value=False), \
              patch("evennia.utils.delay"):
             engine.start_round()
             engine.process_round()
@@ -80,12 +80,12 @@ class TestUnarmedAutoAttack(unittest.TestCase):
 
         with patch("combat.combat_actions.utils.inherits_from", return_value=True), \
              patch("world.system.state_manager.apply_regen"), \
-             patch("combat.combat_actions.check_hit", return_value=(True, "")) as mock_hit, \
-             patch("combat.combat_actions.apply_critical", side_effect=lambda a, t, d: (d, False)), \
-             patch("combat.engine.utils.roll_evade", return_value=False), \
-             patch("combat.engine.utils.roll_parry", return_value=False), \
-             patch("combat.engine.utils.roll_block", return_value=False), \
-             patch("combat.engine.utils.roll_dice_string", return_value=2), \
+             patch("combat.engine.combat_math.CombatMath.check_hit", return_value=(True, "")) as mock_hit, \
+             patch("combat.engine.combat_math.CombatMath.apply_critical", side_effect=lambda a, t, d: (d, False)), \
+             patch("combat.engine.combat_math.roll_evade", return_value=False), \
+             patch("combat.engine.combat_math.roll_parry", return_value=False), \
+             patch("combat.engine.combat_math.roll_block", return_value=False), \
+             patch("combat.engine.combat_math.roll_dice_string", return_value=2), \
              patch("world.system.state_manager.get_effective_stat", return_value=0), \
              patch("evennia.utils.delay"), \
              patch("random.randint", return_value=0):
@@ -107,12 +107,12 @@ class TestUnarmedAutoAttack(unittest.TestCase):
 
         with patch("combat.combat_actions.utils.inherits_from", return_value=True), \
              patch("world.system.state_manager.apply_regen"), \
-             patch("combat.combat_actions.check_hit", return_value=(True, "")) as mock_hit, \
-             patch("combat.combat_actions.apply_critical", side_effect=lambda a, t, d: (d, False)), \
-             patch("combat.engine.utils.roll_evade", return_value=False), \
-             patch("combat.engine.utils.roll_parry", return_value=False), \
-             patch("combat.engine.utils.roll_block", return_value=False), \
-             patch("combat.engine.utils.roll_dice_string", return_value=3), \
+             patch("combat.engine.combat_math.CombatMath.check_hit", return_value=(True, "")) as mock_hit, \
+             patch("combat.engine.combat_math.CombatMath.apply_critical", side_effect=lambda a, t, d: (d, False)), \
+             patch("combat.engine.combat_math.roll_evade", return_value=False), \
+             patch("combat.engine.combat_math.roll_parry", return_value=False), \
+             patch("combat.engine.combat_math.roll_block", return_value=False), \
+             patch("combat.engine.combat_math.roll_dice_string", return_value=3), \
              patch("world.system.state_manager.get_effective_stat", return_value=20), \
              patch("evennia.utils.delay"), \
              patch("random.randint", return_value=0):
